make quic-go build with both Go 1.13 and 1.14
