+++ .travis.yml
@@ -5,6 +5,7 @@ language: go
 
 go:
   - "1.14.x"
+  - "1.13.x"
 
 # first part of the GOARCH workaround
 # setting the GOARCH directly doesn't work, since the value will be overwritten later
@@ -23,6 +24,10 @@ env:
     - TRAVIS_GOARCH=386 TESTMODE=unit
     - TRAVIS_GOARCH=386 TESTMODE=integration
 
+jobs:
+  exclude:
+    - go: "1.13.x"
+      env: TRAVIS_GOARCH=amd64 TESTMODE=fuzz
 
 # second part of the GOARCH workaround
 # now actually set the GOARCH env variable to the value of the temporary variable set earlier
